fix(boot): move runner from api/src/scripts to api/scripts for correct dist path
The CI container-startup job failed with:
  Error: Cannot find module '/app/dist/scripts/run-migrations-with-sentry.js'

`nest build` compiles files at `api/scripts/**` → `api/dist/scripts/**`,
but files at `api/src/scripts/**` land at `api/dist/src/scripts/**`. The
allinone Dockerfile's COPY (`/app/api/dist/scripts → ./dist/scripts`) and
docker-entrypoint's `node /app/dist/scripts/run-migrations-with-sentry.js`
both expect the former. Mirroring the convention of bootstrap-admin.ts,
backfill-game-vectors.ts, etc.

- Move `api/src/scripts/run-migrations-with-sentry.ts` to
  `api/scripts/run-migrations-with-sentry.ts`.
- Update its imports to reach into `../src/...` for sentry/instrument and
  the admin helpers, matching the existing pattern in
  `api/scripts/bootstrap-admin.ts`.
- Keep the unit + integration specs under `api/src/scripts/` (jest's
  rootDir is src/) and update their relative imports to
  `../../scripts/run-migrations-with-sentry`.
- CLAUDE.md note clarifies why the runner lives outside src/.

Local rebuild now produces `api/dist/scripts/run-migrations-with-sentry.js`.
Unit tests still pass (3/3).
