+++ api/src/admin/games-dedup-audit.service.ts
@@ -298,5 +298,3 @@ function buildPersistSummary(
     topGroups,
   };
 }
-
-
